DRAFT: New Select or Shift dropdown and Gamma.
These are all added as hidden comments. They will NOT appear in the public-facing website. I put these in here as a reminder about the feature so that I can save time writing it in. To whom it may concern, please only un-comment these drafted notes when the Select or Shift dropdowns are fully released on the Public Free Poiyomi Shaders package.